Update: Add Fixer method to Linter API
